feat: Create plugin for email notifications
This module allows users to send email notifications upon certain build
events. In their cluster setup (through the Screwdriver API), users can
specify an SMTP host and port, as well as events to subscribe to. When
these events occur (for example, a build success or failure), email
notifications will be sent out to users as described in the
`screwdriver.yaml`s of specific projects.

Co-authored between @jerryzhang222 and @r3rastogi